hack for passing MBC1 tests
Tests that check if CPU reads 0xff from cartridge memory when SRAM is
disabled in MBC fail. Maybe because the pull-up resistors are too
weak. This commit is a temporary fix for that.

Also needed to remove n_emu_mbc pin. Arachne and nextpnr where not able
to route anymore.